pyproject.toml
[tool.poetry]
name = "aws-deploy"
version = "0.1.0"
description = "Full lifecyle deployment from github to aws ecs"
authors = [
    "Ashraful Firoz <ashraf.akon201@gmail.com>",
    "Scotty Rippel <scotty418@gmail.com>"
]
license = "MIT"
readme = "README.md"
packages = [{include = "aws_deploy"}]

[tool.poetry.dependencies]
python = "^3.8.1"
requests = "^2.28.2"
boto3 = "^1.26.120"
cfn-flip = "^1.3.0"
rich = "^13.3.4"
click = "^8.1.3"
pyyaml = "^6.0"
pillow = "^9.5.0"
qrcode = "^7.4.2"
boto3-stubs = "^1.26.120"
mypy-boto3-ssm = "^1.26.97"
mypy-boto3-iam = "^1.26.97"
applescript = "^2021.2.9"
webdriver-manager = "^3.8.6"
docker = "^6.0.1"
pyotp = "^2.8.0"
mypy-boto3-cloudformation = "^1.26.108"
selenium = "^4.9.0"
mypy-boto3-lambda = "^1.26.122"
mypy-boto3-ec2 = "^1.26.122"
mypy-boto3-ecr = "^1.26.0.post1"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
pytest = "^7.3.1"
flake8 = "^6.0.0"


[tool.poetry.scripts]
ad = "aws_deploy.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
